ODE-314 refactor
